requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
pydantic
rapidfuzz
matplotlib
python-multipart
python-docx
pdfplumber
sentence-transformers
jinja2
